requirements.txt
absl-py==2.4.0
basicsr==1.2.0+2b4af71
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
einops==0.8.2
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
grpcio==1.78.0
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
Markdown==3.10.2
MarkupSafe==3.0.2
matplotlib==3.10.8
mpmath==1.3.0
natsort==8.4.0
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.15.1.9
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-python==4.13.0.90
packaging==25.0
pillow==12.1.0
protobuf==6.33.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
PyYAML==6.0.3
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2026.2.16
torch==2.11.0.dev20260118+cu128
torchaudio==2.11.0.dev20260118+cu128
torchvision==0.25.0.dev20260118+cu128
tqdm==4.67.1
triton==3.6.0+git9844da95
typing_extensions==4.15.0
warmup_scheduler==0.3
Werkzeug==3.1.5
wheel==0.45.1
yacs==0.1.8
